refactor(react): extract line and loading components from main view
